Fix heap corruption and verification.
There was a faulty assumption that space End() was always aligned to
card boundaries. This was true for all spaces other than the image
and resulted in heap corruption when the last object of the image
space contained a reference to an object in another space.

Also fixed an error where we called the pre GC verification post GC.

Bug: 17080623

Change-Id: I041ee564518f53b79c6e8dc2ad782a3152577a4e