[Xposed] Add a method to place the hook for a method
This creates a backup of the original method description structure (which is
later used to invoke the original code) and turns the hooked method into a
proxy. The modified proxy invocation handler will take care of routing calls
to XposedBridge.handleHookedMethod().

Some additional information is stored in the field for the native code
pointer.