Don't call AddImageStringsToTable or MoveImageClassesToClassTable if running imageless
If we fail to create the image we use imageless running, this was
breaking AddImageStringsToTable since we didn't check for this
case. Added a check which avoids calling AddImageStringsToTable and
MoveImageClassesToClassTable if we don't have an image.

Bug: 18412472

(cherry picked from commit ab1953f35880f94752e1fcde3f6b3f0d3532125b)

Change-Id: Ida18e1fcd04d752d55923e65c3122c0ac0bde728