ART: ELF symbol table lookup is fixed
ELF loader could not find some symbols in the OAT's symbol table
(e.g. oatdata, oatexec).

There was a bug in the symbol hash table lookup: if a valid index of
a chain got greater or equal than the number of hash buckets then
the chain was cut off.

The fix is to compare the chain index with the chains array length
rather than with the number of buckets.

Change-Id: I20940957c3045913fd426031a51314d4f87ac1bd
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>