ART: reduce UpdateModUnionTable time for partial GC
UpdateModUnionTable is to scan objects with dirty cards in the immune
region. The immune region of partial GC are image space and zygote
space. As image space is always immuned for any kind of GC, there is no
need to scan dirty cards of zygote space pointing to image space for
partial GC.

Change-Id: I2e1f61c2b361985c211076a441418797abf324c2