ART: Materialize constants on environment use
The decision whether to materialize constants assumed that they cannot
be used by an environment. This changed with introduction of the
boolean simplifier and manifested itself when environment uses became
included in liveness analysis. This patch fixes the issue.

This is a quick fix to unblock the build, test will follow.

Bug: 19999189
Change-Id: Iedb1e3a32073a75b9c0d041a88b8ae640e064d13