[Xposed] Don't reuse open oat files from Zygote
Xposed modules are loaded in the Zygote process. ART compiles the .dex file
and opens the resulting .oat file in the Dalvik cache for that. However,
the same happens again when the user installs a new version of the module,
overwriting the file in the Dalvik cache. When opening the new .oat file
though, ART detects that the file is already opened and reuses this instance.
Later, it detects that the checksum doesn't match, which causes a crash.
This will only occur on every second update (without reboot) as Android
uses alternating suffixes for the .apk files (-1.apk / -2.apk).

Let's work around it by ignoring already opened .oat files if they have been
opened in Zygote. The behavior for files in /system (boot image and framework)
is unchanged.

Fixes rovo89/Xposed#22.