[Xposed] Ignore unknown arguments and option values
Some ROMs are configured to make use of options that are only available in
these ROMs, but not in AOSP (and therefore, not in Xposed). There's little
chance that these options can be implemented in Xposed's version of ART as
well. However, these options aren't necessarily required, they might just
be minor tweaks. So we just log them, but don't prevent the device from
booting (unless it happened to be a strictly required option, but then it
wouldn't boot either way).