[Xposed] Don't make original methods private
They need to be handled as direct methods (e.g. no vtable lookups), but
changing the access makes it harder to determine later whether it originally
was a private method already. This distinction is required when retrieving
the oat method index.

Instead, we simulate the direct method flag in IsDirect(), with an optional
parameter to use the original meaning (i.e. handle the method as if it wasn't
modified by Xposed).

For consistency, we add the same parameter to IsProxyMethod() as well.