Move Level1 to its own file.
When we don't have oat files, we use runtime semantics
for checking access. At runtime, we're checking whether
the class loader is the same between two classes. This is
information we don't have when we run dex2oat at install time.

So movining Level1 to a public class fixes the access issue
when running the test with --no-prebuild --no-dex2oat.

Change-Id: Ied1f33a9e651f0045501c939eaf94b7b34ba1e66