Increase alt signal stack to 32K on host.
On host, when there is a crash, the code tries to do an unwind
from the signal handler. Under some conditions this can require
a stack larger than 16K, so increase this value to 32K.

The symptom of this is that you get an empty backtrace running
art tests when there is a crash.

Bug: 21949571
(cherry picked from commit e34c71ce21dbc9fac3427df887e90e1311b6fa8e)

Change-Id: Ib1105a3470064d98843bd46f32d4fadb1d20e416