[Xposed] Support for LG Almond encrypted apps
LG has encrypted a few apps with their DRM implementation. Those apps can
usually not be loaded. Fortunately, the decryption takes place in a separate
library called liblgalmond.so. Xposed now uses this library if available to
check if a dex file (usually located in a ZIP archive) is encrypted, and if
yes, triggers the in-place decryption of the file before initializing it.

Fixes rovo89/Xposed#27.