Optimizing: Improve String.equals() thumb intrinsic.
Use SUBS for updating the loop counter and setting the loop
exit condition, so that we don't need to pre-calculate the
bound but still keep the same number of instructions in the
loop.

This reduces the size of boot.oat on Nexus 5 by 8KiB (when
compiled with Optimizing which is not the default yet).

Change-Id: I87d5a128e5e67f4ad177b71c28662d1367170a10