[Xposed] Read "this" class from debug info for invoke-virtual-quick
Under certain circumstances, the register holding the "this" reference
for an invoke-virtual-quick call might be null. This is usually guarded
by an "if" check in the dex code, but the verifier doesn't analyse the
code in such depth. It aborts because the method cannot be verified,
which makes the whole class unusable.

In such cases, we now check the debug information for type hints for
the register. For now, this is limited to exact matches of the variable
definition range - if really necessary, it could be relaxed to also
considered the closest definition.