[Xposed] Properly compile quickened instructions
This is mostly based on the commit which introduced JIT support in
Marshmallow (e5f13e57ff8fa36342beb33830b3ec5942a61cca), however
many changes needed to be adjusted for Lollipop.

It's not clear to me why this wasn't required before, but on one file,
this solves compilation crashes with compiler filter "everything".