ART: Fix rounding up in DexCacheArraysLayout constructor.
This changes the way we round up to be conceptually correct
but does not really affect the results thanks to current
layout that interleaves GcRoot<> and pointer arrays. If two
odd-length GcRoot<> arrays were followed by a pointer array,
the pointer array would have been previously unaligned on
a 64-bit target.

Change-Id: I6b1cd2ed789f4f91206982caf3c765253fb65824